// src/app/themes-custom/interfaces/interface.ts
export interface IAppConfigFromBackend {
  show_date_filter: any;
  mobile_background_image: string;
  background_image: string;
  app_description: string;
  //header stuff
  web_header_logo: string;
  form_name: string;
  header_color: string;
  color: string;
  is_dual_user_enable: any;
  terminology: ITerminology;

  // is_demo: any;
  // hide_gdpr :any;
  // is_gift_card_activated : any;
  // is_loyalty_point_enabled: any;
  // onboarding_business_type: any;
  // pdp_view: any;
  // languages: any;
  // landing_page_url: string;
  // routes: any;
  // product_view: any;
  // is_homepage_overlay_enabled: number;
  // is_landing_page_enabled: any;
  // signup_field: requiredOnSignup;
  // is_dual_user_enable: number;
  // marketplace_reference_id: string;
  // merchant_domain_obj: IMerchantDomainObj;
  // MERCHANT_COUNT: number;
  // accept_reject_enabled: number;
  // address: string;
  // admin_delivery_method: number;
  // admin_home_delivery: number;
  // admin_self_pickup: number;
  // payment_settings: any;
  // marketplace_user_id: number;
  // android_google_api_key: string;
  // webapp_logo: string;
  // terminology: ITerminology;
  // is_banners_enabled: boolean;
  // app_description: string;
  // app_id: string;
  // app_signup_allowed: number;
  // is_business_category_enabled: boolean;
  // app_url: string;
  // app_version: number;
  // asset_tracking: number;
  // auto_assign: number;
  // autofill_required: number;
  // brand_image: string;
  // btn_color: string;
  // buffer_schedule: number;
  // business_catalog_mapping_enabled: number;
  // business_model_type: string;
  // business_type: number;
  // call_tasks_as: string;
  // cancellation_reason_type: number;
  // category_button_type: number;
  // category_heading: string;
  // category_heading_json: string;
  // category_layout_type: string;
  // category_sub_heading: string;
  // category_sub_heading_json: string;
  // checkout_address_type: number;
  // color: string;
  // company_name: string;
  // contact_details: IContactDetails;
  // contact_email: string;
  // contact_number: string;
  // creation_datetime: string;
  // custom_order_single_screen: number;
  // custom_quotation_enabled: number;
  // custom_service_time: number;
  // dashboard_url: string;
  // days_to_display: number;
  // decimal_calculation_precision_point: number;
  // decimal_display_precision_point: number;
  // default_project_end: number;
  // delivery_by_merchant: number;
  // delivery_charge: string;
  // delivery_method: number;
  // delivery_or_pickup: number;
  // delivery_time: number;
  // display_merchant_address: number;
  // is_review_rating_enabled: boolean;
  // display_merchant_details_page: number;
  // display_merchant_location: number;
  // display_merchant_phone: number;
  // display_slot_intervals: number;
  // distance_range: number;
  // distance_unit: number;
  // dob_on_signup: number;
  // domain_name: string;
  // ds_domain_name: string;
  // duplicate_time_slot: number;
  // dynamic_pages_names: any;
  // dynamic_pages_routes: any;
  // email_service_type: number;
  // enable_custom_tip: number;
  // enable_default_tip: number;
  // enable_merchant_chat: number;
  // enable_merchant_chat_with_marketplace: number;
  // enable_start_time_end_time: number;
  // enable_tookan_agent: number;
  // enable_veg_non_veg_filter: number;
  // enable_veg_non_veg_label: number;
  // enabled_marketplace_storefront: any;
  // enabled_store_slugs: any;
  // facebook_app_id: string;
  // fatafat_res_id: string;
  // fav_icon: string;
  // fav_logo: string;
  // fb_pixel_id: string;
  // fcm_key: string;
  // font: any;
  // force_pickup_delivery: number;
  // form_id: number;
  // form_name: string;
  // form_type: number;
  // fugu_chat_token: string;
  // gift_card_description: string;
  // google_adwords_key: string;
  // google_analytics_client_id: string;
  // google_analytics_is_active: number;
  // google_analytics_tracking_id: string;
  // google_client_app_id: string;
  // google_client_app_ios_id: string;
  // google_client_id: string;
  // google_signature_id: string;
  // has_merchant_fugu_chat_token: number;
  // header_color: string;
  // header_element_color: string;
  // heading: string;
  // heading_json: string;
  // is_google_required: number;
  // is_instagram_required: number;
  // instagram_app_id: string;
  // is_subscription_enabled: number;
  // is_menu_enabled: number;
}

export interface IAppConfig {
  backgroundImage?: string;
  mobileBackgroundImage?: string;
  showDateFilter?: any;
  appDescription?: string;
  webHeaderLogo?: string;
  formName?: string;
  headerColor?: string;
  color?: string;
  isDualUserEnable?: any;
  terminology?: ITerminology;
}

export class ITerminology {
  GIFT_CARD_REDEEEM:string;
  DELIVER_HERE:string;
  CART_READY:string;
  CATALOGUE: string;
  PRODUCT: string;
  CATEGORY: string;
  ORDERS: string;
  ORDER: string;
  MERCHANT: string;
  MERCHANTS: string;
  CUSTOMER: string;
  CUSTOMERS: string;
  STORE: string;
  TOOKAN: string;
  HIPPO: string;
  PICKUP: string;
  DELIVERY: string;
  DELIVERY_CHARGE: string;
  SCHEDULING: string;
  VARIANTS: string;
  INVENTORY: string;
  SERVING: string;
  ITEM: string;
  START_TIME: string;
  END_TIME: string;
  CART: string;
  SCHEDULE_ORDER: string;
  NOTES: string;
  PROCEED_TO_PAY: string;
  CHECKOUT: string;
  PAYMENT_METHOD: string;
  BILL_SUMMARY: string;
  PAY: string;
  PAYMENT: string;
  NOTIFICATIONS: string;
  DELIVERY_FROM: string;
  DELIVER_TO: string;
  BUSINESS: string;
  BUSINESSES: string;
  MENU: string;
  PENDING: string;
  COMPLETED: string;
  REJECTED: string;
  CANCELLED: string;
  ACCEPTED: string;
  WON: string;
  ORDERED: string;
  DISPATCHED: string;
  OPEN: string;
  ONGOING: string;
  ASSIGNED: string;
  PROJECT: string;
  PROJECTS: string;
  EXPIRED: string;
  Ongoing: string;
  CATEGORIES: string;
  PRODUCTS: string;
  CHECKOUT_TEMPLATE: string;
  LOYALTY_POINTS: string;
  UNIT: string;
  DISPATCH: string;
  WALLET: string;
  GIFT_CARD: string;
  BUY: string;
  ITEMS:string;
  REDEEM: string;
  CONTRACTS: string;
  PAY_LATER: string;
  SPONSORED: string;
  SELF_PICKUP: String;
  HOME_DELIVERY: String;
  CANCEL_ORDER:String;
  SEARCH_PRODUCT:String;
  REWARD_CONDITION:String;
  SELF_REDEEM :String;
  REDEEM_GIFT_CARD:String;
  CUSTOMER_RIGHTS:String;

}